# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API Library",
    "description": "Contain all api from kulony backend"
  },
  "servers": [
    {
      "url": "https://kulony-backend.herokuapp.com/"
    }
  ],
  "paths": {
...
    "/api/comment/create": {
      "post": {
        "tags": [
          "Comment"
        ],
        "description": "สร้าง Comment โดยส่ง User_id ของคนคอมเม้น Post_id ของโพสที่จะคอมเม้นและ Comment_content",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "post_id": {
                    "example": "any"
                  },
                  "comment_content": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/comment/{post_id}": {
      "get": {
        "tags": [
          "Comment"
        ],
        "description": "ขอข้อมูลของ Comment ทั้งหมดของ post_id นั้น",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/comment/{entity_id}/report": {
      "post": {
        "tags": [
          "Comment"
        ],
        "description": "ส่ง Report comment",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "report_type": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/comment/{comment_id}/edit": {
      "put": {
        "tags": [
          "Comment"
        ],
        "description": "แก้ไข Comment",
        "parameters": [
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/comment/like/{comment_id}": {
      "put": {
        "tags": [
          "Comment"
        ],
        "description": "Like comment",
        "parameters": [
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/comment/unlike/{comment_id}": {
      "put": {
        "tags": [
          "Comment"
        ],
        "description": "Unlike comment",
        "parameters": [
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/admin/get_all_request_topic": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "รับข้อมูล Request Topic ทั้งหมด",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/remove_request_topic/{request_id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "description": "ลบ Request topic ออกจาก Database",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/accept_request_topic/{request_id}": {
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "ตอบรับ Request topic โดยส่ง topic_name และ catagory_id ที่ต้องการเพิ่มใน Database จากนั้นทำการลบ Request ออกจาก Database",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "catagory_id": {
                    "example": "any"
                  },
                  "topic_name": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}